+++ src/plugins/accordion/accordion.php
@@ -15,13 +15,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
- 
-namespace Phink\Web\UI\Plugin;
+
+namespace Phink\Plugins\Accordion;
 
 use \Phink\Web\UI\Widget\Plugin\TPlugin;
 
 /**
- * Description of newPHPClass
+ * Description of TAccordion
  *
  * @author David
  */
@@ -32,7 +32,7 @@ class TAccordion extends TCustomPlugin
         $result = "\n";
 
         $elements = $this->elements;
-        
+
         $tbody = '';
         $level = 0;
         $boundIndex = 0;
@@ -48,92 +48,75 @@ class TAccordion extends TCustomPlugin
         $bindablesCount = 0;
         $firstBindableIndex = -1;
         $lastBindableIndex = -1;
-        
-        for($k = 0; $k < $templateNum; $k++) {
-            for($j = 0; $j < $this->columns; $j++) {
-                if($this->templates[$k]['name'] == $this->data['names'][$j]) {
+
+        for ($k = 0; $k < $templateNum; $k++) {
+            for ($j = 0; $j < $this->columns; $j++) {
+                if ($this->templates[$k]['name'] == $this->data['names'][$j]) {
                     $this->templates[$k]['index'] = $j;
-                    if($this->templates[$k]['enabled'] == 1) {
+                    if ($this->templates[$k]['enabled'] == 1) {
                         $bindablesCount++;
-                        if($firstBindableIndex == -1) $firstBindableIndex = $j;
+                        if ($firstBindableIndex == -1) {
+                            $firstBindableIndex = $j;
+                        }
+
                         $lastBindableIndex = $j;
                     }
                 }
             }
         }
-                        
-        //self::$logger->debug("\r\n" . "\r\n" . "\r\n" . 'LAST BINDABLE INDEX::' . $lastBindableIndex . "\r\n" . "\r\n" . "\r\n");
-                        
-        for($i = 0; $i < $this->rows; $i++) {
+
+        for ($i = 0; $i < $this->rows; $i++) {
             $row = (isset($body[$i])) ? $body[$i] : array_fill(0, $this->columns, '&nbsp;');
-            for($j = 0; $j < $templateNum; $j++) {
-                 if($j == 0) {
+            for ($j = 0; $j < $templateNum; $j++) {
+                if ($j == 0) {
                     $level = 0;
                     $boundIndex = 0;
                 }
-                if($this->templates[$j]['enabled'] != 1) continue;
+                if ($this->templates[$j]['enabled'] != 1) {
+                    continue;
+                }
+
                 $index = $this->templates[$j]['index'];
                 $canBind = $row[$index] != $oldValue[$j];
-                //$canBind = $canBind && $this->templates[$j]['name'] === $names[$dataIndex];
-                ////self::$logger->debug('TEMPLATE NAME : ' . $this->templates[$j]['name'] . '; HEAD NAME :' . $names[$dataIndex]);
-                if(!$canBind) {
+                if (!$canBind) {
                     $bound[$boundIndex] = $canBind;
-                    //$bound[$boundIndex] = $canBind;
                     $oldLevel = $level;
-                    //$level++;
                     $oldValue[$j] = $row[$index];
                     $boundIndex++;
                     continue;
                 }
-                $level = (($index == $firstBindableIndex) ? 0 : (($index == $lastBindableIndex) ? 2 : 1))    ;  
-                
-                //$html = $row[$index];
-                //$html = $level . '[' . $oldLevel . ']' . '-' . $index . '::' . $row[$index];
+                $level = (($index == $firstBindableIndex) ? 0 : (($index == $lastBindableIndex) ? 2 : 1));
+
                 $dataIndex = array_keys($names, $this->templates[$j]['name'])[0];
                 $noTHead = !empty($this->templates[$j]['content']) && $this->templates[$j]['enabled'] == 1;
 
                 $html = $row[$dataIndex];
-                if($noTHead) {
+                if ($noTHead) {
                     $html = TPlugin::applyTemplate($this->templates, $names, $row, $j);
                 }
 
-                //self::$logger->debug('INDEX::' . $index . "\r\n" . "\r\n");
-                //self::$logger->debug('LEVEL::' . $level . "\r\n" . "\r\n");
-                //self::$logger->debug('HTML::' . $html . "\r\n" . "\r\n");
-                
-                if($level === 0) {
-                    if($i > 0) {
-//                    if($oldLevel === 2) {
-//                        $tbody .= $elements[2]->getClosing() . "\n" . $elements[0]->getClosing() . "\n";
-//                    }
-                        //$bindablesCount = 3;
-                        for($l = 1; $l < $bindablesCount; $l++) {
+                if ($level === 0) {
+                    if ($i > 0) {
+                        for ($l = 1; $l < $bindablesCount; $l++) {
                             $tbody .= $elements[2]->getClosing() . "\n" . $elements[0]->getClosing() . "\n";
-                        } 
+                        }
                         $oldValue = array_fill(0, $this->columns, '!#');
                     }
                     $tbody .= str_replace('%s', 'blue', $elements[0]->getOpening()) . "\n";
                     $tbody .= $elements[1]->getOpening() . $html . $elements[1]->getClosing() . "\n";
                     $tbody .= $elements[2]->getOpening();
-                }
-                elseif($level === 1) { //&& $level < $lastBindableIndex
-                    if($i > 0 && !$bound[$boundIndex - 1]) {
+                } elseif ($level === 1) { //&& $level < $lastBindableIndex
+                    if ($i > 0 && !$bound[$boundIndex - 1]) {
                         $tbody .= $elements[2]->getClosing() . "\n" . $elements[0]->getClosing() . "\n";
-                    } 
-//                    if($oldLevel === 2) {
-//                        $tbody .= $elements[2]->getClosing() . "\n" . $elements[0]->getClosing() . "\n";
-//                    }
+                    }
                     $tbody .= str_replace('%s', 'odd', $elements[0]->getOpening()) . "\n";
                     $tbody .= $elements[1]->getOpening() . $html . $elements[1]->getClosing() . "\n";
                     $tbody .= $elements[2]->getOpening();
-                }
-                elseif($level === 2) {
+                } elseif ($level === 2) {
                     $tbody .= str_replace('%s', '', $elements[2]->getOpening()) . $html . $elements[2]->getClosing() . "\n";
-                }                
+                }
                 $bound[$boundIndex] = $canBind;
-                //$bound[$boundIndex] = $canBind;
                 $oldLevel = $level;
-                //$level++;
                 $boundIndex++;
                 $oldValue[$j] = $row[$index];
             }
@@ -144,7 +127,7 @@ class TAccordion extends TCustomPlugin
         $tbody .= $elements[0]->getClosing() . "\n";
 
         $result .= $tbody;
-        
+
         return $result;
     }
 }
